Tweaked the balancing to be better
