--- Assets/Game/Scripts/Gameplay/BalancingArrow/BalancingInput.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9de7046be52899e4183d68cbdea9c0cb
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
